ContentNode instances now responsible for loading their own content
…giving the chance for subclasses to override the get_content method for their own means. 

Speaking of subclasses, the visitor pattern implementation for the content hierarchy now find the most specific visitXXX() method based on the entire class hierarchy rather than just looking at the leaf class name. This means that it no longer falls back to the visit() method when subclassing a ContentNode or DirectoryNode, which would break generation.